configure: Minimal riscovite OS support
This is just enough to get the configure script to recognize "riscovite" as
a valid OS and to cause executable programs to be named with a "!" suffix
to match the convention for directly-executable programs on RISCovite.

This is sufficient to get the libraries to build. We have no threads
support right now, and also some gaps in our C library, so this can't
build any of the programs. Even if it could, we'd need to add support for
the RISCovite BGAI API before it'd actually be able to play back audio
or video on that platform.